requirements.txt
google-cloud-run==0.5.0
google-cloud-artifact-registry==1.8.1
docker==6.1.2
PyJWT==2.3.0
google-auth==2.22.0
flask==2.0.1
werkzeug==2.0.3
gunicorn==20.1.0
